[CodeClean] fix svace issue
1. change data type to handle gst-memory.
2. use util function to get nth tensor info.
3. add null-init in filter-edgetpu.
4. fix use-after-free case when parsing accel in tensor-filter.

+++ ext/nnstreamer/tensor_converter/tensor_converter_flatbuf.cc
@@ -73,7 +73,7 @@ fbc_convert (GstBuffer *in_buf, GstTensorsConfig *config, void *priv_data)
   GstBuffer *out_buf = NULL;
   GstMemory *in_mem, *out_mem;
   GstMapInfo in_info;
-  guint mem_size;
+  gsize mem_size;
   GstTensorInfo *_info;
 
   UNUSED (priv_data);
